Re-arrange file structure for later improved routing
