+++ app/(app)/dashboard_contributor.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import config from "../tamagui.config";
+import config from "../../tamagui.config";
 import { Session } from "@supabase/supabase-js";
 import {
   YStack,
@@ -13,7 +13,7 @@ import {
   ScrollView,
   XStack,
 } from "tamagui";
-import LessonCard from "../src/components/LessonCard";
+import LessonCard from "../../src/components/LessonCard";
 
 export default function ContributorDashboard({
   session,
@@ -38,26 +38,36 @@ export default function ContributorDashboard({
         <View marginStart="$2" marginTop="$3" marginEnd="$2">
           <Card elevate size={4} bordered height={200}>
             <CardHeader padded>
-              <Text color={"$color"} fontSize={26}>Contributor Dashboard</Text>
+              <Text color={"$color"} fontSize={26}>
+                Contributor Dashboard
+              </Text>
               <XStack marginTop={"$3"}>
-                <Text color={"$color"} fontSize={"$7"} flex={1}>Total Contribution: </Text>
-                <Text color={"$color"} fontSize={"$7"}>50</Text>
+                <Text color={"$color"} fontSize={"$7"} flex={1}>
+                  Total Contribution:{" "}
+                </Text>
+                <Text color={"$color"} fontSize={"$7"}>
+                  50
+                </Text>
               </XStack>
               <XStack marginTop={"$3"}>
-                <Text color={"$color"} fontSize={"$7"} flex={1}>Total Views: </Text>
-                <Text color={"$color"} fontSize={"$7"}>50</Text>
+                <Text color={"$color"} fontSize={"$7"} flex={1}>
+                  Total Views:{" "}
+                </Text>
+                <Text color={"$color"} fontSize={"$7"}>
+                  50
+                </Text>
               </XStack>
               <XStack marginTop={"$3"}>
-                <Text color={"$color"} fontSize={"$7"} flex={1}>Total Likes: </Text>
-                <Text color={"$color"} fontSize={"$7"}>50</Text>
+                <Text color={"$color"} fontSize={"$7"} flex={1}>
+                  Total Likes:{" "}
+                </Text>
+                <Text color={"$color"} fontSize={"$7"}>
+                  50
+                </Text>
               </XStack>
             </CardHeader>
-            <CardFooter padded>
-
-            </CardFooter>
-            <CardBackground>
-
-            </CardBackground>
+            <CardFooter padded></CardFooter>
+            <CardBackground></CardBackground>
           </Card>
         </View>
         <View marginStart="$2" marginTop="$5" marginEnd="$2">
