+++ app/(app)/dictionary.tsx
@@ -15,15 +15,15 @@ import {
   Spinner,
 } from "tamagui";
 import React, { useState, useEffect } from "react";
-import { supabase } from "../src/utils/supabase";
+import { supabase } from "../../src/utils/supabase";
 import { Alert } from "react-native";
-import { Word } from "../src/models/Word";
-import AddWordDialog from "../src/components/AddWordDialog";
-import { WordSearchResult } from "../src/components/WordSearchResult";
-import { DictionaryService } from "../src/services/DictionaryService";
+import { Word } from "../../src/models/Word";
+import AddWordDialog from "../../src/components/AddWordDialog";
+import { WordSearchResult } from "../../src/components/WordSearchResult";
+import { DictionaryService } from "../../src/services/DictionaryService";
 import { Link } from "expo-router";
-import EditWordDialog from "../src/components/EditWordDialog";
-import EditTranslationDialog from "../src/components/EditTranslationDialog";
+import EditWordDialog from "../../src/components/EditWordDialog";
+import EditTranslationDialog from "../../src/components/EditTranslationDialog";
 
 export default function Dictionary({
   session,
