+++ app/(app)/exercises/grammar/[exercise_id].tsx
@@ -22,11 +22,11 @@ import { useEffect, useState } from "react";
 import { Alert, useColorScheme } from "react-native";
 import { ExerciseService } from ".../../../src/services/ExerciseService";
 import { useLocalSearchParams } from "expo-router";
-import { VocabularyExercise } from "../../../src/models/VocabularyExercise";
-import { VocabularyExerciseUI } from "../../../src/components/ExerciseUI";
-import { VocabularyExerciseType } from "../../../src/utils/enums";
+import { GrammarExerciseUI } from "../../../../src/components/ExerciseUI";
+import { GrammarExerciseType } from "../../../../src/utils/enums";
+import { GrammarExercise } from "../../../../src/models/GrammarExercise";
 
-export default function VocabularyExercises({
+export default function GrammarExercises({
   session,
   exercise_id,
 }: {
@@ -35,18 +35,14 @@ export default function VocabularyExercises({
 }) {
   // DO NOT DELETE: FOR TESTING AND INITIALIZATION
   useEffect(() => {
-    console.log("VOCABULARY_EXERCISES_" + local.exercise_id + " page loaded.");
+    console.log("GRAMMAR_EXERCISES_" + local.exercise_id + " page loaded.");
   }, []);
 
-  const TEMP_USER_UUID = "ebabaa6c-4254-465e-9f2f-f285a2364277";
   const colorScheme = useColorScheme();
   const local = useLocalSearchParams();
   const [loading, setLoading] = useState(false);
   const [loadingText, setLoadingText] = useState("Loading...");
-  const [exercise, setExercise] = useState<VocabularyExercise | null>();
-  const [vocabExerType, setVocabExerType] = useState<VocabularyExerciseType>(
-    VocabularyExerciseType.ChooseCebRepresentationForEngWord
-  );
+  const [exercise, setExercise] = useState<GrammarExercise | null>();
 
   useEffect(() => {
     loadExercise();
@@ -56,17 +52,10 @@ export default function VocabularyExercises({
     try {
       setLoadingText("Loading exercise...");
       setLoading(true);
-      let problems = await ExerciseService.getVocabularyExerciseProblems(
+      let problems = await ExerciseService.getGrammarExerciseProblems(
         parseInt(local.exercise_id as string)
       );
-      let exerType: VocabularyExerciseType =
-        await ExerciseService.getVocabularyExerciseType(
-          parseInt(local.exercise_id as string),
-          TEMP_USER_UUID
-        );
-      setVocabExerType(exerType);
       setExercise(problems);
-      console.log(("EXERTPE IS CURRENTLY " + vocabExerType) as string);
       console.log("*****" + problems);
     } catch (error) {
       if (error instanceof Error) {
@@ -81,7 +70,7 @@ export default function VocabularyExercises({
     <TamaguiProvider>
       <Theme name={colorScheme === "dark" ? "dark" : "light"}>
         <YStack f={1} jc="center" ai="stretch" backgroundColor={"$background"}>
-          {loading && exercise ? (
+          {loading ? (
             <YStack jc="flex-start" ai="center" padding="$5">
               <Spinner size="large" color="$blue9" m="$2" />
               <Text fontSize={20} fontWeight={400} color={"$color"}>
@@ -89,9 +78,8 @@ export default function VocabularyExercises({
               </Text>
             </YStack>
           ) : (
-            <VocabularyExerciseUI
-              exercise_id={parseInt(local.exercise_id as string)}
-              exercise_type={vocabExerType}
+            <GrammarExerciseUI
+              exercise_type={GrammarExerciseType.InputEnglishSentence}
               exercise={exercise || null}
             />
           )}
